+++ client/src/css/AppStyles.module.css
@@ -1,0 +1,276 @@
+/* ------------------------------
+   Additional Shared Styles
+------------------------------ */
+.container {
+    display: flex;
+    min-height: 100vh;
+    position: relative;
+  }
+  
+  .content {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    min-height: 100vh;
+  }
+  
+  .header {
+    background: #555;
+    color: white;
+    padding: 30px;
+    text-align: center;
+  }
+  
+  .middle {
+    flex: 1;
+    padding: 20px;
+    background: #f4f4f4;
+    overflow: auto;
+  }
+  
+  .inputBox {
+    margin-bottom: 5px;
+    display: block;
+    padding: 8px;
+    width: 50%;
+  }
+
+  .submitButton {
+    margin-top: 5px;
+    padding: 10px;
+    background-color: blue;
+    color: white;
+    border: none;
+    cursor: pointer;
+    width: 20%;
+  }
+  
+  .submitButton:hover {
+    background-color: darkblue;
+  }
+  
+  .login-prompt-overlay {
+    position: fixed;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background-color: rgba(0, 0, 0, 0.5);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    z-index: 1000;
+  }
+  
+  .login-prompt-content {
+    background: white;
+    padding: 20px;
+    border-radius: 8px;
+    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
+    text-align: center;
+    max-width: 400px;
+    width: 90%;
+  }
+  
+  .login-prompt-content h2 {
+    margin-bottom: 15px;
+    color: #333;
+  }
+  
+  .login-prompt-content p {
+    color: #666;
+    margin-bottom: 20px;
+  }
+  
+  .login-prompt-content button {
+    padding: 8px 16px;
+    background-color: #007bff;
+    color: white;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+  }
+  
+  .login-prompt-content button:hover {
+    background-color: #0056b3;
+  }
+  
+  /*---------------------
+       PopUps
+---------------------*/ 
+.popupOverlay {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background-color: rgba(0,0,0,0.5);
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+  
+  .popup {
+    background: white;
+    padding: 20px;
+    border-radius: 12px;
+    width: 400px;
+    max-height: 80vh;
+    overflow-y: auto;
+  }
+
+  .popupImg {
+    width: 100%;
+    height: auto;
+    margin-bottom: 10px;
+    border-radius: 12px;
+    display: block;
+  }
+
+  /*----------------------
+  Popup design
+-----------------------*/
+
+.popup-overlay {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background: rgba(0, 0, 0, 0.5);
+    z-index: 1000;
+  }
+  
+  .popup-box {
+    position: absolute;
+    top: 20%;
+    left: 50%;
+    transform: translate(-50%, 0);
+    background: white;
+    padding: 20px;
+    border-radius: 12px;
+    box-shadow: 0 0 10px rgba(0, 0, 0, 0.3);
+  }
+  /* ----------------
+  Admin menu design
+-------------------*/
+.adminMenuContainer {
+    display: inline-block;
+    margin-left: 20px;
+    position: relative;
+  }
+  
+  .adminMenuToggle {
+    background: none;
+    border: none;
+    cursor: pointer;
+    font: inherit;
+    color: #61dafb;
+  }
+  
+  .adminMenuDropdown {
+    position: absolute;
+    top: 100%;
+    left: 0;
+    background-color: white;
+    border: 1px solid #ccc;
+    padding: 10px;
+    z-index: 1000;
+    min-width: 180px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+    display: flex;
+    flex-direction: column;
+  }
+  
+  .adminMenuDropdown a {
+    padding: 5px 10px;
+    text-decoration: none;
+    color: #333;
+    border-radius: 4px;
+  }
+  
+  .adminMenuDropdown a:hover {
+    background-color: #f0f0f0;
+  }
+
+/*----------------------
+  Map layers menu design
+-----------------------*/
+.layerPanelWrapper {
+    position: absolute;
+    top: 10px;
+    left: 10px;
+    z-index: 1000;
+  }
+  
+  .layerToggleButton {
+    background-color: #f0f0f0;
+    border: 1px solid #ccc;
+    border-radius: 6px;
+    padding: 6px 10px;
+    cursor: pointer;
+    font-size: 14px;
+    margin-bottom: 5px;
+  }
+  
+  .layerTogglePanel {
+    background-color: rgba(255, 255, 255, 0.95);
+    padding: 10px;
+    border-radius: 8px;
+    box-shadow: 0 2px 6px rgba(0, 0, 0, 0.2);
+    font-size: 14px;
+    max-height: 200px;
+    overflow-y: auto;
+  }
+  
+  .legend {
+    display: flex; /* Align items horizontally */
+    flex-wrap: wrap; /* Allow wrapping to the next line if needed */
+    gap: 3px;
+    background: white;
+    padding: 3px;
+    border-radius: 4px;
+    box-shadow: 0 0 4px rgba(0,0,0,0.3);
+    position: absolute;
+    top: 130px;
+    left: 20px;
+    z-index: 1000;
+  }
+  .legendItem {
+    display: flex;
+    align-items: center;
+    margin-bottom: 5px;
+    color: black; /* Ensure the text color is visible */
+    font-size: 14px;
+  }
+
+  /*-----------------------
+  Map design
+------------------------*/
+.recenterButton {
+    position: absolute;
+    top: 100px;
+    right: 20px;
+    z-index: 1000;
+    width: 20px; /* Set the width of the button */
+    height: 20px; /* Set the height of the button */
+    background-color: white;
+    border: 1px solid gray;
+    border-radius: 50%; /* Make it circular */
+    cursor: pointer;
+    background-image: url('../icons/crosshair2.png'); /* Path to your image */
+    background-size: cover; /* Ensure the image covers the button */
+    background-position: center; /* Center the image */
+    background-repeat: no-repeat; /* Prevent tiling */
+    transition: transform 0.2s;
+  }
+  
+  .recenterButton:hover {
+    transform: scale(1.1); /* Slight zoom effect on hover */
+  }
+
+  .mapControls {
+    display: flex;
+    gap: 10px;
+    margin-bottom: 10px;
+  }
